Split network mod.rs into multiple files
